Move TLS syntax to a separable library
