+++ lib/tls_syntax/test/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(TEST_APP_NAME "${LIB_NAME}_test")
+set(TEST_APP_NAME "${CURRENT_LIB_NAME}_test")
 
 # Dependencies
 find_package(doctest REQUIRED)
@@ -7,8 +7,8 @@ find_package(doctest REQUIRED)
 file(GLOB TEST_SOURCES CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
 
 add_executable(${TEST_APP_NAME} ${TEST_SOURCES})
-add_dependencies(${TEST_APP_NAME} ${LIB_NAME} tls_syntax)
-target_link_libraries(${TEST_APP_NAME} ${LIB_NAME} tls_syntax doctest::doctest OpenSSL::Crypto)
+add_dependencies(${TEST_APP_NAME} ${CURRENT_LIB_NAME})
+target_link_libraries(${TEST_APP_NAME} ${CURRENT_LIB_NAME} doctest::doctest)
 
 # Enable CTest
 include(doctest)
