+++ src/tree_math.cpp
@@ -1,6 +1,5 @@
 #include "mls/tree_math.h"
 #include "mls/common.h"
-#include "mls/tls_syntax.h"
 
 #include <algorithm>
