+++ test/messages.cpp
@@ -1,7 +1,7 @@
 #include "test_vectors.h"
 #include <doctest/doctest.h>
 #include <mls/messages.h>
-#include <mls/tls_syntax.h>
+#include <tls/tls_syntax.h>
 
 using namespace mls;
