+++ test/test_vectors.h
@@ -2,7 +2,7 @@
 #include <mls/messages.h>
 #include <mls/session.h>
 #include <mls/state.h>
-#include <mls/tls_syntax.h>
+#include <tls/tls_syntax.h>
 #include <mls/tree_math.h>
 #include <string>
